feat: [FE] 장바구니 가져오는 함수 utils로 분리
