+++ packages/client/src/utils/index.ts
@@ -1,6 +1,3 @@
-import { CART_KEY } from '@/constants';
-import { CartMenu } from 'types/Cart';
-
 /**
  * 금액 세자리 단위 콤마 추가
  */
@@ -13,18 +10,4 @@ export const getPriceComma = (price: number | string) => {
  */
 export const getFirstUpper = (text: string) => {
   return `${text.slice(0, 1).toUpperCase()}${text.slice(1)}`;
-};
-
-/**
- * 장바구니 담긴 갯수 return
- */
-export const getCartCount = () => {
-  const cart = JSON.parse(localStorage.getItem(CART_KEY) || '[]');
-  let count = 0;
-  
-  cart.forEach((menu: CartMenu) => {
-    count += menu.quantity;
-  });
-  
-  return count;
 };
